requirements.txt
ccxt>=4.4.82
fastapi[standard]>=0.115.12
uvicorn[standard]>=0.34.2
numpy>=1.26.0
python-dotenv>=1.1.0
aiohttp>=3.10.10
python-telegram-bot>=22.1